+++ src/Config.hs
@@ -230,6 +230,7 @@ sourceFile Doraemon = "doraemon"
 sourceFile FinalSpace = "final_space"
 sourceFile HowToTrainYourDragon = "how_to_train_your_dragon"
 sourceFile Mountain = "mountain"
+sourceFile Naruto = "naruto"
 
 mapSource :: Text -> Name
 mapSource "studioGhibli" = 'StudioGhibli
@@ -409,6 +410,7 @@ mapSource "finalSpace" = 'FinalSpace
 mapSource "howToTrainYourDragon" = 'HowToTrainYourDragon
 mapSource "jackHandey" = 'JackHandey
 mapSource "mountain" = 'Mountain
+mapSource "naruto" = 'Naruto
 mapSource item = error $ "mapSource: Invalid argument passed " <> (show item)
 
 guessSourceFile :: SourceData -> Text -> FilePath